requirements.txt
numpy==1.19.5
pandas==1.3.4
Pillow==8.4.0
base64
os
streamlit_folium
branca
